src/requirements.txt
Django==4.2.9
django-environ==0.11.2
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
django-phonenumber-field==7.3.0
phonenumbers==8.13.27
psycopg2-binary==2.9.9
Pillow==10.2.0
requests==2.31.0
rich==13.5.2
torch==2.2.2
websockets==12.0
idna==3.7
protobuf>=3.12,<4
streamlit==1.22.0
